% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_tbl.R
\name{augment_tbl}
\alias{augment_tbl}
\title{Augment a table with additional columns.}
\usage{
augment_tbl(tbl, conn, with_info, on_col = "AuthorId", ...)
}
\arguments{
\item{tbl}{A query from \code{conn} with dbplyr and lazily evaluated.}

\item{conn}{An object of class \code{SQLiteConnection} to a sqlite database.}

\item{with_info}{Which info should \code{tbl} be augmented with?
A column vector with the following options, specified as strings:
\itemize{
\item affiliation: joins information on unit-year for units in \code{on_col}
\item output: joins research output on unit-year for units in \code{on_col}
\item coauthor: joins the (academic lifetime) co-authors of units in \code{on_col}
}}

\item{on_col}{On which column should the information be joined? The default
is "AuthorId", the unit of authors in MAG. Alternatively, use "CoAuthorId" to
join information on co-authors (see below for details).}

\item{...}{Additional arguments to be passed to
\code{\link{make_tbl_output}}.
If not specified, a lazily evaluated table without limit is returned.
Partially specified arguments are completed with
\code{\link{dots_tbl_output}}.}
}
\value{
A new \code{tbl} with the columns specified \code{with_info} added.
}
\description{
This function adds additional data to records in a table: data on
the affiliation, publication output and co-authors of records in the
table.
}
\details{
There are two main purposes for which this function can be currently
used:
\enumerate{
\item Join output and/or affiliation information to
author units. This works directly with one call to \code{augment_tbl}.
\item Join information on affiliations of co-authors
of author units in \code{tbl}. To do this, you need to call \code{augment_tbl} twice:
First, to join the co-author information of author units in \code{tbl}, and then
again to join the affiliations of co-authors,
using the option \verb{on_col = "CoAuthorId}.
}

Mixing the purposes is discouraged because it creates duplicated records.
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
graduates <- get_links(conn, from = "graduates")
graduates <- augment_tbl(graduates, conn, with_info = "output")

}
