% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors_proquest.R
\name{authors_proquest}
\alias{authors_proquest}
\title{Source PhD graduates}
\usage{
authors_proquest(conn, start_year = 1985, end_year = 2005, ...)
}
\arguments{
\item{conn}{An object of the DBIConnection class.}

\item{start_year}{Lowest graduation year to consider. Default: 1985.}

\item{end_year}{Highest graduation year to consider. Default: 2005.}

\item{...}{additional arguments to be passed on to be passed on to
\code{\link{make_tbl_output}}}
}
\value{
A table with U.S. PhD graduates and their gender.
}
\description{
Source PhD graduates
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
d_graduates <- authors_proquest(conn = conn)

}
