% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to the database}
\usage{
connect_to_db(db_file)
}
\arguments{
\item{db_file}{full name (including path) of the file.}
}
\value{
An object of class \code{SQLiteConnection} to a sqlite database.
}
\description{
A wrapper around \code{\link[DBI]{dbConnect}} for Sqlite.
Create an object of class \code{SQLiteConnection} to a sqlite database stored in
\code{db_file}. Prints the output from \code{\link[dbplyr]{src_dbi}} and returns
the connection object.
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
}
