% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_helpers.R
\name{copy_fixture}
\alias{copy_fixture}
\title{Copy fixtures to make available for tests}
\usage{
copy_fixture(
  mock_db = "./tests/testthat/mock_db/",
  origin,
  filename,
  verbose = FALSE
)
}
\arguments{
\item{mock_db}{The path to the mock database containing the fixtures for
dittodb. Default is \verb{tests/testthat/mock_db/}.}

\item{origin}{The origin directory where dittodb saved the fixtures
from capturing}

\item{filename}{The name of the file to be copied.}

\item{verbose}{Logical. Should filenames be printed?
The filename consist of the first SQL verb of the query
(often "SELECT") and the
hash of the query that is sent.
See the dittodb documentation for details.}
}
\value{
Copies the fixture from \verb{mock_db/origin/} to \verb{mock_db/}.
}
\description{
A wrapper to around \code{\link[base]{dput}} to copy a database fixture
from \code{origin} to a directory with fixtures used by dittodb.
}
\keyword{internal}
