% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{db_example}
\alias{db_example}
\title{Generate the path to the example database}
\usage{
db_example(file)
}
\arguments{
\item{file}{Name of sqlite database for examples.}
}
\value{
A string with the path to the database.
}
\description{
Make it easy to access the database for examples.
}
\examples{
db_example("AcademicGraph.sqlite")
}
