% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_field.R
\name{define_field}
\alias{define_field}
\title{Define the field of study}
\usage{
define_field(tbl, conn, from, ...)
}
\arguments{
\item{tbl}{A query from \code{conn} with dbplyr and lazily evaluated.}

\item{conn}{An object of class \code{SQLiteConnection} to a sqlite database.}

\item{from}{A string with options to be queried: "mag_authors" or "graduates".}

\item{...}{Additional arguments to be passed to
\code{\link{make_tbl_output}}.
If not specified, a lazily evaluated table without limit is returned.
Partially specified arguments are completed with
\code{\link{dots_tbl_output}}.}
}
\value{
\code{tbl} augmented by a column with the name of the field of study at
level 0 as defined by MAG.
}
\description{
This function defines the field of study at level 0 in MAG of records
in \code{tbl}. To do so, it queries tables in the database that store the
field of study of the person. See the details for more information.
}
\details{
\subsection{When using \code{from} = "graduates"}{

The field is defined with a custom mapping between the reported field of
study in ProQuest and the fields in MAG. The mapping is stored in the
database.
A record can have multiple fields, and currently the function returns the
first reported. A missing \code{fieldname0_mag} indicates that it is not possible
to map the field at position 0 in ProQuest to the MAG fields.
In future, this may be made more flexible to consider any of the reported
fields in ProQuest.
}

\subsection{When using \code{from} = "mag_authors"}{

Returns the one field for which the person is observed the most likely to
publish over their career. The field is defined based on a confidence score
of the likely field of study across all publications in the career.
}
}
\keyword{internal}
