% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-master.R
\name{doc_sqlite_connection}
\alias{doc_sqlite_connection}
\title{Collect documentation of commonly used function input}
\usage{
doc_sqlite_connection(conn)
}
\arguments{
\item{conn}{An object of class \code{SQLiteConnection} to a sqlite database.}
}
\description{
Collect documentation of commonly used function input
}
\details{
The function does noting. It just collects all the
arguments that are repeatedly used in various functions throughout the
package for easy reference and maintenance.
}
