% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dots_tbl_output}
\alias{dots_tbl_output}
\title{Handle dots for making table output}
\usage{
dots_tbl_output(...)
}
\arguments{
\item{...}{Ellipsis, passed on from a higher function}
}
\value{
If \code{...} is empty, returns NULL. Otherwise it converts
\code{...} into a named list with elements "lazy" and "limit".
}
\description{
This is a helper function to transform ... from user-facing functions
into arguments for \code{\link{make_tbl_output}}.
}
\details{
For non-empty but partially specified arguments, it adds
the following defaults: \code{lazy = TRUE}, \code{limit = Inf}.
For instance, if only limit = 3 is passed, \code{\link{make_tbl_output}}
will return a lazily evaluated query.
To safeguard against accidentally loading large queries into memory
by only specifying \code{lazy = FALSE}, an error is thrown in this case.
}
\keyword{internal}
