% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environment.R
\name{get_linked_graduates}
\alias{get_linked_graduates}
\title{Load links between MAG and ProQuest}
\usage{
get_linked_graduates(conn, keep_unique = TRUE)
}
\arguments{
\item{conn}{A DBI connection.}

\item{keep_unique}{If TRUE (the default), drops graduates that have multiple links to MAG.}
}
\value{
A lazy query of linked goid-AuthorId.
}
\description{
Load links between MAG and ProQuest
}
