% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Load links between MAG and ProQuest}
\usage{
get_links(conn, from, min_score = 0.7, ...)
}
\arguments{
\item{conn}{An object of the DBIConnection class to a database.}

\item{from}{The table with the links to be used.
Must be "advisors" or "graduates"}

\item{min_score}{Minimum score for links to accept. Numeric between 0 and 1.}

\item{...}{additional arguments to be passed on to be passed on to
\code{\link{make_tbl_output}}.
If not specified, a lazily evaluated table without limit is returned.
Partially specified arguments are completed with \code{\link{dots_tbl_output}}.}
}
\value{
A query of linked goid-AuthorId.
}
\description{
Load links between MAG and ProQuest
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
links <- get_links(conn, from = "graduates", min_score = 0.7)

}
