% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{graduate_fields}
\alias{graduate_fields}
\title{Define the field of study for graduates.}
\usage{
graduate_fields(conn, lazy = TRUE, limit = Inf)
}
\arguments{
\item{conn}{An object of the DBIConnection class.}

\item{lazy}{If TRUE (the default), does not \code{collect()} the query into a dataframe.
This is useful if other tables from the database are joined later on.}

\item{limit}{LIMIT of the query. A positive integer or Inf.
Default is Inf, in which case all records are returned.}
}
\value{
A table with one field name for each goid.
The field name is the name of the field at level 0 in MAG. The field is
defined with a custom mapping between the reported field of study in ProQuest
and the fields in MAG. The mapping is stored in the database.

\strong{Note}: A record can have multiple fields, and currently the function returns the
first reported. A missing \code{fieldname0_mag} indicates that it is not possible
to map the field at position 0 in ProQuest to the MAG fields.
In future, this may be made more flexible to consider any of the reported
fields in ProQuest.
}
\description{
Define the field of study for graduates.
}
\examples{
\dontrun{
d <- graduate_fields(db_example("AcademicGraph.sqlite"))
}
}
