% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_tbl_output}
\alias{make_tbl_output}
\title{Make table output from a SQL query}
\usage{
make_tbl_output(tbl, limit, lazy)
}
\arguments{
\item{tbl}{A query from \code{conn} with dbplyr and lazily evaluated.}

\item{limit}{LIMIT of the query. A positive integer or Inf.
Default is Inf, in which case all records are returned.}

\item{lazy}{If TRUE, does not \code{collect()} the query into a dataframe.
This is useful if other tables from the database are joined later on.}
}
\value{
A query, evaluated with \code{limit} and \code{lazy}.
}
\description{
This function is used at the end of user-facing functions to query the
database. It limits the query to some number of rows and collects if
if specified.
}
\keyword{internal}
