% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_checking.R
\name{tidy_string}
\alias{tidy_string}
\title{Remove redundant white spaces and new lines}
\usage{
tidy_string(s)
}
\arguments{
\item{s}{A string.}
}
\value{
The string, with new lines and excess white spaces removed.
Excess white spaces refer to both trailining and leading as well as
multiple white spaces next to each other.
}
\description{
Remove redundant white spaces and new lines
}
\examples{
magutils:::tidy_string("  some string     with  many   white spaces. ")
}
