% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{authors_proquest}
\alias{authors_proquest}
\title{Source PhD graduates}
\usage{
authors_proquest(conn, start_year = 1985, end_year = 2005)
}
\arguments{
\item{conn}{A DBI connection.}

\item{start_year}{Lowest graduation year to consider. Default: 1985.}

\item{end_year}{Highest graduation year to consider. Default: 2005.}
}
\value{
A lazily evaluated table with U.S. PhD graduates and their gender.
}
\description{
Source PhD graduates
}
