% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{authors_proquest}
\alias{authors_proquest}
\title{Source PhD graduates}
\usage{
authors_proquest(
  conn,
  start_year = 1985,
  end_year = 2005,
  lazy = TRUE,
  limit = Inf
)
}
\arguments{
\item{conn}{An object of the DBIConnection class.}

\item{start_year}{Lowest graduation year to consider. Default: 1985.}

\item{end_year}{Highest graduation year to consider. Default: 2005.}

\item{lazy}{If TRUE (the default), does not \code{collect()} the query into a dataframe.
This is useful if other tables from the database are joined later on.}

\item{limit}{LIMIT of the query. A positive integer or Inf.
Default is Inf, in which case all records are returned.}
}
\value{
A table with U.S. PhD graduates and their gender.
}
\description{
Source PhD graduates
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
d_graduates <- authors_proquest(conn = conn)

}
