% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environment.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to the database}
\usage{
connect_to_db(db_file)
}
\arguments{
\item{db_file}{full name (including path) to the file}
}
\value{
A DBI::dbConnect object
}
\description{
Connect to the database
}
