% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_field.R
\name{define_field}
\alias{define_field}
\title{Define the field of study for records in a table.}
\usage{
define_field(tbl, conn, from, ...)
}
\arguments{
\item{tbl}{A tbl on which to add the field of study.}

\item{conn}{An object of the DBIConnection class.}

\item{from}{A string with options to be queried: "mag_authors" or "graduates".}

\item{...}{additional arguments to be passed on to be passed on to
\code{\link{make_tbl_output}}.
If not specified, a lazily evaluated table without limit is returned.
Partially specified arguments are completed with \code{\link{dots_tbl_output}}.}
}
\value{
A table with one field name for each person id.
The field name is the name of the field at level 0 in MAG.
The person id is \code{AuthorId} for \code{from} = "mag_authors" and
\code{goid} for \code{from} = "graduates".
}
\description{
Define the field of study for records in a table.
}
\details{
\subsection{When using \code{from} = "graduates"}{

The field is defined with a custom mapping between the reported field of
study in ProQuest and the fields in MAG. The mapping is stored in the
database.
A record can have multiple fields, and currently the function returns the
first reported. A missing \code{fieldname0_mag} indicates that it is not possible
to map the field at position 0 in ProQuest to the MAG fields.
In future, this may be made more flexible to consider any of the reported
fields in ProQuest.
}

\subsection{When using \code{from} = "mag_authors"}{

Returns the one field for which the person is observed the most likely to
publish over their career. The field is defined based on a confidence score
of the likely field of study across all publications in the career.
}
}
