% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{define_gender}
\alias{define_gender}
\title{Define gender based on first name.}
\usage{
define_gender(conn, table, firstname_left, drop_missing)
}
\arguments{
\item{conn}{A DBI connection.}

\item{table}{A lazily evaluated table sourced from \code{conn}.}

\item{firstname_left}{Column containing the firstname in \code{table} and to be used for joining gender on.}

\item{drop_missing}{If TRUE, drops records without clear gender assigned.
Clear assignment is when probability of either gender is >= 0.8}
}
\value{
\code{table} augmented by a gender column.
}
\description{
Define gender based on first name.
}
\examples{
\dontrun{
new_table <- define_gender(
conn = conn, table = old_table,
firstname_left = "firstname_old", drop_missing = TRUE
)
}

}
