% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environment.R
\name{define_gender}
\alias{define_gender}
\title{Define gender based on first name.}
\usage{
define_gender(conn, table, drop_missing)
}
\arguments{
\item{conn}{A DBI connection.}

\item{table}{A lazily evaluated table sourced from \code{conn}.}

\item{drop_missing}{If TRUE, drops records without clear gender assigned.}
}
\description{
Define gender based on first name.
}
