% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dots_tbl_output}
\alias{dots_tbl_output}
\title{Helper function to handle dots for making table output}
\usage{
dots_tbl_output(...)
}
\arguments{
\item{...}{Ellipsis, passed on from a higher function}
}
\value{
The ellipsis as a list, with defaults for \code{make_tbl_output} added:
\code{lazy} is by default \code{TRUE}, \code{limit} is by default \code{Inf}. Thus, if you only
pass on limit = 3, you get a lazily evaluated query.
To safeguard against accidentally loading large queries into memory
by only specifying \code{lazy = FALSE}, an error is thrown in this case.
}
\description{
Helper function to handle dots for making table output
}
