% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_graduate_links}
\alias{get_graduate_links}
\title{Load links between MAG and ProQuest}
\usage{
get_graduate_links(conn, keep_unique = TRUE, limit = Inf, lazy = TRUE)
}
\arguments{
\item{conn}{A DBI connection.}

\item{keep_unique}{If TRUE (the default), drops graduates that have multiple links to MAG.}

\item{limit}{LIMIT of the query. A positive integer or Inf.
Default is Inf, in which case all records are returned.}

\item{lazy}{If TRUE (the default), does not \code{collect()} the query into a dataframe.
This is useful if other tables from the database are joined later on.}
}
\value{
A query of linked goid-AuthorId.
}
\description{
Load links between MAG and ProQuest
}
