% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_graduate_links}
\alias{get_graduate_links}
\title{Load links between MAG and ProQuest}
\usage{
get_graduate_links(conn, keep_unique = TRUE)
}
\arguments{
\item{conn}{A DBI connection.}

\item{keep_unique}{If TRUE (the default), drops graduates that have multiple links to MAG.}
}
\value{
A lazy query of linked goid-AuthorId.
}
\description{
Load links between MAG and ProQuest
}
