% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Load links between MAG and ProQuest}
\usage{
get_links(conn, from, min_score = 0.7, limit = Inf, lazy = TRUE)
}
\arguments{
\item{conn}{An object of the DBIConnection class to a database.}

\item{from}{The table with the links to be used.
Must be "advisors" or "graduates"}

\item{min_score}{Minimum score for links to accept. Numeric between 0 and 1.}

\item{limit}{LIMIT of the query. A positive integer or Inf.
Default is Inf, in which case all records are returned.}

\item{lazy}{If TRUE (the default), does not \code{collect()} the query into a dataframe.
This is useful if other tables from the database are joined later on.}
}
\value{
A query of linked goid-AuthorId.
}
\description{
Load links between MAG and ProQuest
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
links <- get_links(conn, from = "graduates", min_score = 0.7)

}
