% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Load links between MAG and ProQuest}
\usage{
get_links(conn, from, min_score = 0.7, ...)
}
\arguments{
\item{conn}{An object of class \code{SQLiteConnection} to a sqlite database.}

\item{from}{A string with options to be queried: "advisors" or "graduates".}

\item{min_score}{Minimum score for links to accept. Numeric between 0 and 1.}

\item{...}{Additional arguments to be passed to
\code{\link{make_tbl_output}}.
If not specified, a lazily evaluated table without limit is returned.
Partially specified arguments are completed with
\code{\link{dots_tbl_output}}.}
}
\value{
A query of linked goid-AuthorId.
}
\description{
Load the links between records in MAG and in ProQuest. The links are
stored as a table in the database. The links can be either between
PhD graduates and MAG authors, or PhD advisors and MAG authors.
}
\details{
For advisor links, a high threshold for \code{min_score} (0.95 or higher)
is recommended, and the function throws a message if that is not the case.
The reason for this is that there seem to be many false positives when using
a lower threshold.
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
graduate_links <- get_links(conn, from = "graduates", min_score = 0.7)
advisor_links <- get_links(conn, from = "advisors", min_score = 0.7)

}
