% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proquest.R
\name{get_proquest}
\alias{get_proquest}
\title{Load ProQuest data}
\usage{
get_proquest(conn, from, start_year = 1985, end_year = 2005, ...)
}
\arguments{
\item{conn}{An object of class \code{SQLiteConnection} to a sqlite database.}

\item{from}{A string with options to be queried: "advisors" or "graduates".}

\item{start_year}{Lowest graduation year to consider. Default: 1985.}

\item{end_year}{Highest graduation year to consider. Default: 2005.}

\item{...}{Additional arguments to be passed to
\code{\link{make_tbl_output}}.
If not specified, a lazily evaluated table without limit is returned.
Partially specified arguments are completed with
\code{\link{dots_tbl_output}}.}
}
\value{
For graduates, returns a table with degree year, university id,
fieldname and gender.
For advisors, returns a table with gender and relationship id.
}
\description{
This function loads data on PhD graduates or their advisors
from ProQuest. It is based on metadata from ProQuest,
and automatically adds the likely gender of the person,
using \code{\link{define_gender}}, and for graduates, their
field using \code{\link{define_field}}.
}
\details{
For simplicity, does not return the degree year and university information to
advisors. You can add this information to advisors by joining the
output of \code{get_proquest(from = "graduates")} to the output of
\code{get_proquest(from = "advisors")}.
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
d_graduates <- get_proquest(conn = conn, from = "graduates")
d_advisors <- get_proquest(conn = conn, from = "advisors")

}
