% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_checking.R
\name{get_tbl_idx}
\alias{get_tbl_idx}
\title{Extract the table indexes from a table}
\usage{
get_tbl_idx(conn, tbl, temp = FALSE)
}
\arguments{
\item{conn}{A DBI connection to a sqlite database.}

\item{tbl}{The name of the table.}

\item{temp}{Should \code{sqlite_temp_master} be queried, instead of \code{sqlite_master}?
Default is FALSE. This can be useful when looking for temporary tables and
indexes on them.}
}
\value{
A named list of lists. Each list corresponds to one index on \code{tbl}.
A elements (top-level) of the list are named according to the name of the
indexes in the database.
Each element is a list with two entries:
\itemize{
\item \code{idx_unique}: A logical indicating whether the index is unique.
\item \code{idx_cols}: A character vector with the index columns.
}
}
\description{
Extract the table indexes from a table
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
get_tbl_idx(conn, "author_output")
}
