% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_tbl_output}
\alias{make_tbl_output}
\title{Make final table output from a sql query}
\usage{
make_tbl_output(tbl, limit, lazy)
}
\arguments{
\item{tbl}{A lazily evaluated query from sql with dbplyr.}

\item{limit}{LIMIT of the query. A positive integer or Inf.
Default is Inf, in which case all records are returned.}

\item{lazy}{If TRUE, does not \code{collect()} the query into a dataframe.
This is useful if other tables from the database are joined later on.}
}
\value{
A query, evaluated with \code{limit} and \code{lazy}.
}
\description{
Make final table output from a sql query
}
