% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{names_tbl_lazy}
\alias{names_tbl_lazy}
\title{Extract the names of a lazily evaluated table}
\usage{
names_tbl_lazy(tbl)
}
\arguments{
\item{tbl}{A lazily evaluated table from dplyr.}
}
\value{
A vector with the names of the dataframe if \code{tbl} was \code{collect}ed.
}
\description{
Extract the names of a lazily evaluated table
}
