% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_checking.R
\name{sqlite_master_to_df}
\alias{sqlite_master_to_df}
\title{Transform sqlite_master table to a dataframe.}
\usage{
sqlite_master_to_df(conn)
}
\arguments{
\item{conn}{An object of the DBIConnection class.}
}
\value{
A dataframe with type, name, tbl_name and sql statement
from sqlite_master.
}
\description{
Transform sqlite_master table to a dataframe.
}
\examples{
conn <- connect_to_db(db_example("AcademicGraph.sqlite"))
sqlite_master_to_df(conn)
}
